+++ libft_last_ver/libft.h
@@ -67,11 +67,11 @@ char				*ft_strsub(char const *s, unsigned int start, size_t len);
 char				*ft_strjoin(char const *s1, char const *s2);
 char				*ft_strtrim(char const *s);
 char				**ft_strsplit(char const *s, char c);
-char				*ft_itoa(long long int n);
+char				*ft_itoa(intmax_t n);
 void				ft_putchar(char c);
 void				ft_putstr(char const *s);
 void				ft_putendl(char const *s);
-void				ft_putnbr(int n);
+void				ft_putnbr(intmax_t n);
 void				ft_putchar_fd(char c, int fd);
 void				ft_putstr_fd(char const *s, int fd);
 void				ft_putendl_fd(char const *s, int fd);
@@ -84,8 +84,8 @@ void				ft_lstiter(t_list *lst, void (*f)(t_list *elem));
 size_t				ft_strcl(const char *s, char c);
 char				*ft_strnjoin(char const *s1, char const *s2);
 char				*ft_strnjoinfree(char *s1, char const *s2);
-void				ft_itoo(char *str, int n);
-void				ft_itox(char *str, int n);
+void				ft_itoo(char *str, intmax_t n);
+void				ft_itox(char *str, intmax_t n);
 int					ft_printf(const char *restrict format, ...);
 
 #endif
